gpu: Remove gpuav_error_message.cpp
All functions in there were dedicated to log error messages coming
from gpuav shader instrumentation, so moved everything to
gpuav_instrumentation
