+++ layers/CMakeLists.txt
@@ -242,7 +242,6 @@ target_sources(vvl PRIVATE
     gpu/core/gpuav_setup.cpp
     gpu/debug_printf/debug_printf.cpp
     gpu/debug_printf/debug_printf.h
-    gpu/error_message/gpuav_error_message.cpp
     gpu/error_message/gpuav_vuids.cpp
     gpu/error_message/gpuav_vuids.h
     gpu/instrumentation/gpu_shader_instrumentor.cpp
